Changes made according review
Modified table sort function, seperated table function from utils and removed redundant import statements
