+++ index.html
@@ -17,8 +17,6 @@
             </div>  
         </header>
 
-        
-
         <!--main div class="container"-->
         <main class="container" id="main">
             <div class="content">
@@ -61,6 +59,7 @@
         <script type="module" src="scripts/utils.js"></script>
         <script type="module" src="scripts/menu.js"></script>
         <script type="module" src="scripts/home.js"></script>
+        <script type="module" src="scripts/table.js"></script>
         <script type="module" src="scripts/latestPosts.js"></script>
     </body>
 </html>
