+++ services.html
@@ -65,6 +65,7 @@
     <script type="module" src="scripts/utils.js"></script>
     <script type="module" src="scripts/menu.js"></script>
     <script type="module" src="scripts/services.js"></script>
+    <script type="module" src="scripts/table.js"></script>
     <script type="module" src="scripts/latestPosts.js"></script>
 </body>
 </html>
